# openapi/wordquest-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WordQuest API",
    "description": "词境英语的服务端 API。当前阶段提供基础设施与系统接口。",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer" } },
...
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "summary": "注册学习账户",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "displayName", "deviceId"],
                "properties": {
                  "email": { "type": "string", "format": "email", "maxLength": 320 },
                  "password": { "type": "string", "minLength": 8, "maxLength": 128 },
                  "displayName": { "type": "string", "minLength": 1, "maxLength": 100 },
                  "deviceId": { "type": "string", "minLength": 1, "maxLength": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["token", "expiresAt", "user"],
                  "properties": {
                    "token": { "type": "string" },
                    "expiresAt": { "type": "string", "format": "date-time" },
                    "user": {
                      "type": "object",
                      "required": ["id", "email", "displayName"],
                      "properties": {
                        "id": { "type": "string" },
                        "email": { "type": "string" },
                        "displayName": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "timestamp", "path"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "statusCode", "requestId"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "statusCode": { "type": "integer" },
                        "requestId": { "type": "string" },
                        "details": {}
                      }
                    },
                    "timestamp": { "type": "string", "format": "date-time" },
                    "path": { "type": "string" }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "timestamp", "path"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "statusCode", "requestId"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "statusCode": { "type": "integer" },
                        "requestId": { "type": "string" },
                        "details": {}
                      }
                    },
                    "timestamp": { "type": "string", "format": "date-time" },
                    "path": { "type": "string" }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "timestamp", "path"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "statusCode", "requestId"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "statusCode": { "type": "integer" },
                        "requestId": { "type": "string" },
                        "details": {}
                      }
                    },
                    "timestamp": { "type": "string", "format": "date-time" },
                    "path": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "登录账户",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "deviceId"],
                "properties": {
                  "email": { "type": "string", "format": "email", "maxLength": 320 },
                  "password": { "type": "string", "minLength": 8, "maxLength": 128 },
                  "displayName": { "type": "string", "minLength": 1, "maxLength": 100 },
                  "deviceId": { "type": "string", "minLength": 1, "maxLength": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["token", "expiresAt", "user"],
                  "properties": {
                    "token": { "type": "string" },
                    "expiresAt": { "type": "string", "format": "date-time" },
                    "user": {
                      "type": "object",
                      "required": ["id", "email", "displayName"],
                      "properties": {
                        "id": { "type": "string" },
                        "email": { "type": "string" },
                        "displayName": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "timestamp", "path"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "statusCode", "requestId"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "statusCode": { "type": "integer" },
                        "requestId": { "type": "string" },
                        "details": {}
                      }
                    },
                    "timestamp": { "type": "string", "format": "date-time" },
                    "path": { "type": "string" }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "timestamp", "path"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "statusCode", "requestId"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "statusCode": { "type": "integer" },
                        "requestId": { "type": "string" },
                        "details": {}
                      }
                    },
                    "timestamp": { "type": "string", "format": "date-time" },
                    "path": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "summary": "获取当前账户",
        "tags": ["auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id", "email", "displayName"],
                  "properties": {
                    "id": { "type": "string" },
                    "email": { "type": "string" },
                    "displayName": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "timestamp", "path"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "statusCode", "requestId"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "statusCode": { "type": "integer" },
                        "requestId": { "type": "string" },
                        "details": {}
                      }
                    },
                    "timestamp": { "type": "string", "format": "date-time" },
                    "path": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "退出当前账户",
        "tags": ["auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Default Response" },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error", "timestamp", "path"],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": ["code", "message", "statusCode", "requestId"],
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" },
                        "statusCode": { "type": "integer" },
                        "requestId": { "type": "string" },
                        "details": {}
                      }
                    },
                    "timestamp": { "type": "string", "format": "date-time" },
                    "path": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [{ "url": "http://localhost:3001", "description": "development" }],
...
}